+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+## Unreleased
+
+* Fixed `react export` issue that caused the process to hang indefinitely
+* Added changelog/contributing files to server template
+* Updated server template's ci.yaml GitHub workflow with multiple fixes
+* Removed unused files from react template
+* Updated react and server template package dependencies to latest versions
+* The `dev` command now sets `NODE_ENV` to `development`
+
 ## v1.0.1
 
 * Fixed react template to add missing manifestPath to config.ts